Fixing bug in matrix_multiply.py
generator is created on cpu by default so trying to create a tensor on
cuda with a cpu generator would cause an error, we can just move it over
instead.

added a test to demonstrate that this works and was broken before.
added random-matrix to the other correctness tests

(test was generated with claude-code and verified by me)

Signed-off-by: HDCharles <charlesdavidhernandez@gmail.com>